The current state of the project
I felt the files were very scattered in the HardwareAssembler and I am having some difficulty committing changes into directories.  Here is the zip of the current project with all the updates.  The only name change is that the script "search_git.sh+ was changed to "repo_clone_test.sh" since I added the function of running the testbench and validating hardware components when they are cloned.